+++ R/driver-sql-server.R
@@ -81,6 +81,36 @@ setMethod("dbExistsTable", c("Microsoft SQL Server", "character"),
   }
 )
 
+#' @description
+#' ## `dbListTables()`
+#' The default implementation reports temporary tables as non-existent
+#' when a `catalog_name` isn't supplied since they live in a different catalog.
+#' This method provides a special case for temporary tables.
+#' @rdname SQLServer
+#' @usage NULL
+setMethod("dbListTables", "Microsoft SQL Server",
+  function(conn,
+           catalog_name = NULL,
+           schema_name = NULL,
+           table_name = NULL,
+           table_type = NULL,
+           ...) {
+    res <- callNextMethod()
+
+    if (is.null(catalog_name) && is.null(schema_name)) {
+      res_temp <- callNextMethod(
+        conn = conn,
+        catalog_name = "tempdb",
+        schema_name = "dbo"
+      )
+
+      res <- c(res, res_temp)
+    }
+
+    res
+  }
+)
+
 #' @rdname SQLServer
 #' @usage NULL
 setMethod("dbExistsTable", c("Microsoft SQL Server", "Id"),
@@ -104,6 +134,7 @@ setMethod("dbExistsTable", c("Microsoft SQL Server", "SQL"),
 
 #' @description
 #' ## `odbcConnectionSchemas`
+#'
 #' Method for an internal function. Calls catalog-specific `sp_tables` to make
 #' sure we list the schemas in the appropriate database/catalog.
 #' @rdname SQLServer